chores: code clean up

+++ Allay-API/src/main/java/cn/allay/api/math/vector/ImplLoc3l.java
@@ -265,7 +265,7 @@ record ImplLoc3l(long x, long y, long z, double pitch, double yaw, double headYa
 
     @Override
     public double distance(final long x, final long y, final long z) {
-        return (double) Math.sqrt(this.distanceSquared(x, y, z));
+        return Math.sqrt(this.distanceSquared(x, y, z));
     }
 
     @Override
@@ -275,7 +275,7 @@ record ImplLoc3l(long x, long y, long z, double pitch, double yaw, double headYa
 
     @Override
     public double length() {
-        return (double) Math.sqrt(this.lengthSquared());
+        return Math.sqrt(this.lengthSquared());
     }
 
     /**
