+++ Allay-API/src/main/java/cn/allay/api/math/vector/ImplMutablePos3d.java
@@ -57,7 +57,7 @@ final class ImplMutablePos3d implements MutablePos3d {
     }
 
     public ImplMutablePos3d(final float x, final float y, final float z, final World world) {
-        this((double) x, (double) y, (double) z, world);
+        this(x, y, (double) z, world);
     }
 
     public ImplMutablePos3d(final double x, final double y, final double z, final World world) {
@@ -159,7 +159,7 @@ final class ImplMutablePos3d implements MutablePos3d {
 
     @Override
     public MutablePos3d add(final float x, final float y, final float z) {
-        return this.add((double) x, (double) y, (double) z);
+        return this.add(x, y, (double) z);
     }
 
     @Override
@@ -174,7 +174,7 @@ final class ImplMutablePos3d implements MutablePos3d {
 
     @Override
     public MutablePos3d sub(final float x, final float y, final float z) {
-        return this.sub((double) x, (double) y, (double) z);
+        return this.sub(x, y, (double) z);
     }
 
     @Override
@@ -199,7 +199,7 @@ final class ImplMutablePos3d implements MutablePos3d {
 
     @Override
     public MutablePos3d mul(final float x, final float y, final float z) {
-        return this.mul((double) x, (double) y, (double) z);
+        return this.mul(x, y, (double) z);
     }
 
     @Override
@@ -224,7 +224,7 @@ final class ImplMutablePos3d implements MutablePos3d {
 
     @Override
     public MutablePos3d div(final float x, final float y, final float z) {
-        return this.div((double) x, (double) y, (double) z);
+        return this.div(x, y, (double) z);
     }
 
     @Override
@@ -239,7 +239,7 @@ final class ImplMutablePos3d implements MutablePos3d {
 
     @Override
     public double dot(final float x, final float y, final float z) {
-        return this.dot((double) x, (double) y, (double) z);
+        return this.dot(x, y, (double) z);
     }
 
     @Override
@@ -254,7 +254,7 @@ final class ImplMutablePos3d implements MutablePos3d {
 
     @Override
     public MutablePos3d project(final float x, final float y, final float z) {
-        return this.project((double) x, (double) y, (double) z);
+        return this.project(x, y, (double) z);
     }
 
     @Override
@@ -274,7 +274,7 @@ final class ImplMutablePos3d implements MutablePos3d {
 
     @Override
     public MutablePos3d cross(final float x, final float y, final float z) {
-        return this.cross((double) x, (double) y, (double) z);
+        return this.cross(x, y, (double) z);
     }
 
     @Override
@@ -324,7 +324,7 @@ final class ImplMutablePos3d implements MutablePos3d {
 
     @Override
     public MutablePos3d min(final float x, final float y, final float z) {
-        return this.min((double) x, (double) y, (double) z);
+        return this.min(x, y, (double) z);
     }
 
     @Override
@@ -339,7 +339,7 @@ final class ImplMutablePos3d implements MutablePos3d {
 
     @Override
     public MutablePos3d max(final float x, final float y, final float z) {
-        return this.max((double) x, (double) y, (double) z);
+        return this.max(x, y, (double) z);
     }
 
     @Override
@@ -354,7 +354,7 @@ final class ImplMutablePos3d implements MutablePos3d {
 
     @Override
     public double distanceSquared(final float x, final float y, final float z) {
-        return this.distanceSquared((double) x, (double) y, (double) z);
+        return this.distanceSquared(x, y, (double) z);
     }
 
     @Override
@@ -372,12 +372,12 @@ final class ImplMutablePos3d implements MutablePos3d {
 
     @Override
     public double distance(final float x, final float y, final float z) {
-        return this.distance((double) x, (double) y, (double) z);
+        return this.distance(x, y, (double) z);
     }
 
     @Override
     public double distance(final double x, final double y, final double z) {
-        return (double) Math.sqrt(this.distanceSquared(x, y, z));
+        return Math.sqrt(this.distanceSquared(x, y, z));
     }
 
     @Override
@@ -387,7 +387,7 @@ final class ImplMutablePos3d implements MutablePos3d {
 
     @Override
     public double length() {
-        return (double) Math.sqrt(this.lengthSquared());
+        return Math.sqrt(this.lengthSquared());
     }
 
     @Override
