+++ Allay-API/src/main/java/cn/allay/api/math/vector/ImplMutablePos3l.java
@@ -332,7 +332,7 @@ final class ImplMutablePos3l implements MutablePos3l {
 
     @Override
     public double distance(final long x, final long y, final long z) {
-        return (double) Math.sqrt(this.distanceSquared(x, y, z));
+        return Math.sqrt(this.distanceSquared(x, y, z));
     }
 
     @Override
@@ -342,7 +342,7 @@ final class ImplMutablePos3l implements MutablePos3l {
 
     @Override
     public double length() {
-        return (double) Math.sqrt(this.lengthSquared());
+        return Math.sqrt(this.lengthSquared());
     }
 
     /**
