+++ Allay-API/src/main/java/cn/allay/api/math/vector/ImplPos3d.java
@@ -52,7 +52,7 @@ record ImplPos3d(double x, double y, double z, World world) implements Pos3d {
     }
 
     public ImplPos3d(final float x, final float y, final float z, final World world) {
-        this((double) x, (double) y, (double) z, world);
+        this(x, y, (double) z, world);
     }
 
     @Override
@@ -77,7 +77,7 @@ record ImplPos3d(double x, double y, double z, World world) implements Pos3d {
 
     @Override
     public Pos3d add(final float x, final float y, final float z) {
-        return this.add((double) x, (double) y, (double) z);
+        return this.add(x, y, (double) z);
     }
 
     @Override
@@ -92,7 +92,7 @@ record ImplPos3d(double x, double y, double z, World world) implements Pos3d {
 
     @Override
     public Pos3d sub(final float x, final float y, final float z) {
-        return this.sub((double) x, (double) y, (double) z);
+        return this.sub(x, y, (double) z);
     }
 
     @Override
@@ -117,7 +117,7 @@ record ImplPos3d(double x, double y, double z, World world) implements Pos3d {
 
     @Override
     public Pos3d mul(final float x, final float y, final float z) {
-        return this.mul((double) x, (double) y, (double) z);
+        return this.mul(x, y, (double) z);
     }
 
     @Override
@@ -142,7 +142,7 @@ record ImplPos3d(double x, double y, double z, World world) implements Pos3d {
 
     @Override
     public Pos3d div(final float x, final float y, final float z) {
-        return this.div((double) x, (double) y, (double) z);
+        return this.div(x, y, (double) z);
     }
 
     @Override
@@ -157,7 +157,7 @@ record ImplPos3d(double x, double y, double z, World world) implements Pos3d {
 
     @Override
     public double dot(final float x, final float y, final float z) {
-        return this.dot((double) x, (double) y, (double) z);
+        return this.dot(x, y, (double) z);
     }
 
     @Override
@@ -172,7 +172,7 @@ record ImplPos3d(double x, double y, double z, World world) implements Pos3d {
 
     @Override
     public Pos3d project(final float x, final float y, final float z) {
-        return this.project((double) x, (double) y, (double) z);
+        return this.project(x, y, (double) z);
     }
 
     @Override
@@ -192,7 +192,7 @@ record ImplPos3d(double x, double y, double z, World world) implements Pos3d {
 
     @Override
     public Pos3d cross(final float x, final float y, final float z) {
-        return this.cross((double) x, (double) y, (double) z);
+        return this.cross(x, y, (double) z);
     }
 
     @Override
@@ -242,7 +242,7 @@ record ImplPos3d(double x, double y, double z, World world) implements Pos3d {
 
     @Override
     public Pos3d min(final float x, final float y, final float z) {
-        return this.min((double) x, (double) y, (double) z);
+        return this.min(x, y, (double) z);
     }
 
     @Override
@@ -257,7 +257,7 @@ record ImplPos3d(double x, double y, double z, World world) implements Pos3d {
 
     @Override
     public Pos3d max(final float x, final float y, final float z) {
-        return this.max((double) x, (double) y, (double) z);
+        return this.max(x, y, (double) z);
     }
 
     @Override
@@ -272,7 +272,7 @@ record ImplPos3d(double x, double y, double z, World world) implements Pos3d {
 
     @Override
     public double distanceSquared(final float x, final float y, final float z) {
-        return this.distanceSquared((double) x, (double) y, (double) z);
+        return this.distanceSquared(x, y, (double) z);
     }
 
     @Override
@@ -290,12 +290,12 @@ record ImplPos3d(double x, double y, double z, World world) implements Pos3d {
 
     @Override
     public double distance(final float x, final float y, final float z) {
-        return this.distance((double) x, (double) y, (double) z);
+        return this.distance(x, y, (double) z);
     }
 
     @Override
     public double distance(final double x, final double y, final double z) {
-        return (double) Math.sqrt(this.distanceSquared(x, y, z));
+        return Math.sqrt(this.distanceSquared(x, y, z));
     }
 
     @Override
@@ -305,7 +305,7 @@ record ImplPos3d(double x, double y, double z, World world) implements Pos3d {
 
     @Override
     public double length() {
-        return (double) Math.sqrt(this.lengthSquared());
+        return Math.sqrt(this.lengthSquared());
     }
 
     @Override
