+++ Allay-API/src/main/java/cn/allay/api/math/vector/ImplPos3l.java
@@ -266,7 +266,7 @@ record ImplPos3l(long x, long y, long z, World world) implements Pos3l {
 
     @Override
     public double distance(final long x, final long y, final long z) {
-        return (double) Math.sqrt(this.distanceSquared(x, y, z));
+        return Math.sqrt(this.distanceSquared(x, y, z));
     }
 
     @Override
@@ -276,7 +276,7 @@ record ImplPos3l(long x, long y, long z, World world) implements Pos3l {
 
     @Override
     public double length() {
-        return (double) Math.sqrt(this.lengthSquared());
+        return Math.sqrt(this.lengthSquared());
     }
 
     /**
